SubSampledImage-12 Add option to disable gestures
