+++ sample/src/main/java/xyz/quaver/graphics/subsampledimage/sample/MainActivity.kt
@@ -19,18 +19,16 @@ package xyz.quaver.graphics.subsampledimage.sample
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.produceState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowUp
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -43,35 +41,58 @@ class MainActivity : ComponentActivity() {
 
     val logger = newLogger(LoggerFactory.default)
 
+    @ExperimentalMaterialApi
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
+            val imageSource = resources.assets.open("card.png").use { rememberInputStreamImageSource(it) }
+
+            val states = remember {
+                List(5) { _ ->
+                    SubSampledImageState(ScaleTypes.FIT_WIDTH, Bounds.FORCE_OVERLAP_OR_CENTER)
+                }
+            }
+
+            var isGestureEnabled by remember { mutableStateOf(false) }
+
+            LaunchedEffect(isGestureEnabled) {
+                states.forEach {
+                    it.isGestureEnabled = isGestureEnabled
+                }
+            }
+
             SubSamplingImageViewTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(color = MaterialTheme.colors.background) {
-                    val imageSource = resources.assets.open("card.png").use { rememberInputStreamImageSource(it) }
-
-                    val states = listOf(
-                        rememberSubSampledImageState(ScaleTypes.FIT_WIDTH),
-                        rememberSubSampledImageState(ScaleTypes.FIT_WIDTH),
-                        rememberSubSampledImageState(ScaleTypes.FIT_WIDTH),
-                        rememberSubSampledImageState(ScaleTypes.FIT_WIDTH),
-                        rememberSubSampledImageState(ScaleTypes.FIT_WIDTH),
-                    )
-
+                BottomSheetScaffold(
+                    sheetContent = {
+                        Card(Modifier.fillMaxWidth(), elevation = 4.dp) {
+                            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                                Icon(Icons.Default.KeyboardArrowUp, contentDescription = null, modifier = Modifier.size(32.dp))
+                                Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
+                                    Text("Enable Gestures")
+                                    Switch(isGestureEnabled, onCheckedChange = {
+                                        isGestureEnabled = it
+                                    })
+                                }
+                            }
+                        }
+                    }, sheetPeekHeight = 32.dp
+                ) {
                     LazyColumn(modifier = Modifier.fillMaxSize()) {
                         items(states) { state ->
                             val height by produceState<Float?>(null, state.canvasSize, state.imageSize) {
                                 if (value != null) return@produceState
 
                                 state.canvasSize?.let { canvasSize ->
-                                state.imageSize?.let { imageSize ->
-                                    value = imageSize.height * canvasSize.width / imageSize.width
-                                } }
+                                    state.imageSize?.let { imageSize ->
+                                        value = imageSize.height * canvasSize.width / imageSize.width
+                                    } }
                             }
 
                             SubSampledImage(
-                                modifier = Modifier.height(height?.let { with(LocalDensity.current) { it.toDp() } } ?: 128.dp).fillMaxWidth(),
+                                modifier = Modifier
+                                    .height(height?.let { with(LocalDensity.current) { it.toDp() } }
+                                        ?: 128.dp)
+                                    .fillMaxWidth(),
                                 imageSource = imageSource,
                                 state = state)
                         }
